requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# Database
motor>=3.3.2
pymongo>=4.6.0
beanie>=1.30.0

# NLP and ML
spacy>=3.7.2
nltk>=3.8.1
scikit-learn>=1.3.2
numpy>=1.26.1
pandas>=2.1.3

# Transformers and advanced NLP (install separately for better compatibility)
transformers>=4.35.0
# torch will be installed separately via PyTorch index
sentencepiece>=0.1.99

# Visualization
wordcloud>=1.9.3
matplotlib>=3.8.2
plotly>=5.17.0
seaborn>=0.13.0
Pillow>=10.1.0

# Web and utilities
python-dotenv>=1.0.0
aiofiles>=23.2.1
httpx>=0.25.2
jinja2>=3.1.2

# Development and testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
isort>=5.12.0

# API documentation
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Additional NLP libraries
sumy>=0.11.0
textblob>=0.17.1

# Additional dependencies for full functionality
reportlab>=4.0.0
openpyxl>=3.1.0
python-docx>=0.8.11
pdfplumber>=0.11.4
pdfminer.six>=20231228
fasttext>=0.9.2
langdetect>=1.0.9
lime>=0.2.0.1
shap>=0.42.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
vaderSentiment==3.3.2

# Optional: For advanced Indian language support
# indic-nlp-library==0.92
# indic-transliterate==2.3.57

# CORS and security
python-multipart==0.0.6